Rename ParternsItemLayout to PartnersIcon
