+++ cs-expo-2023/components/PartnersIcon.tsx
